rename crash models
these models crashed the game when they were first released. They were distributed to a lot of players on the twlz server, and the fixed versions were not renamed. So, I'm renaming them in case anyone still has the old versions, and blocking use of the old versions on twlz.